etl/pyproject.toml
[project]
name = "etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.11",
    "aioredis>=2.0.1",
    "asyncpg>=0.30.0",
    "pydantic>=2.10.5",
    "python-dotenv>=1.0.1",
    "pytz>=2024.2",
    "redis>=5.2.1",
]
